requirements.txt
joblib
keras
nltk
numpy
scikit-learn
streamlit
tensorflow-cpu